pyproject.toml
[project]
name = "notecast"
version = "0.1.0"
dependencies = [
    "podgen>=1.0.0",
    "requests>=2.31.0",
    "notebooklm-py>=0.3.4",
    "playwright>=1.40.0",
    "aiohttp>=3.9.0",
    "feedparser>=6.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["notecast"]

